--- src/header/Header.jsx
@@ -1,26 +1,0 @@
-import styled from 'styled-components';
-import React from 'react';
-
-function Header() {
-  return (
-    <StHeader>
-      <StText>TMIL</StText>
-    </StHeader>
-  );
-}
-
-const StHeader = styled.div`
-  height: 80px;
-  width: 1920px;
-  padding-left: 20px;
-  font-size: 30px;
-  background-color: #c9eff9;
-  display: table-cell;
-  vertical-align: middle;
-`;
-
-const StText = styled.p`
-  margin: 0px;
-`;
-
-export default Header;
